+++ src/app.cc
@@ -29,8 +29,8 @@ int App::process_event(int event_type, int param_one, int param_two) {
   if (event_type != EVT_MTSYNC) {
     // No need to log multi touch synchronization events
 
-    BOOST_LOG_TRIVIAL(debug) << "Processing event of type "
-                             << format_event_type(event_type);
+    BOOST_LOG_TRIVIAL(debug)
+        << "Processing event of type " << format_event_type(event_type);
   }
   if (event_type == EVT_INIT) {
     this->setup();